pyproject.toml
[project]
name = "donna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
]
